+++ incar.py
@@ -1,5 +1,8 @@
 from pathlib import Path
+from typing import Union
+import numpy as np
 from vasprunio import read_vasprun
+import yaml
 
 class Incar:
 
@@ -14,9 +17,11 @@ class Incar:
         '''Creates a dictionary of the INCAR file from vasprun.xml'''
         incar_children = self.root.find('incar').getchildren()
         incar_dict = {child.attrib['name']                      : child.text for child in incar_children}
-        # trim leading and trailing whitespace from the values
         incar_dict = {k: v.strip() for k, v in incar_dict.items()}
-
+        #remove KPOINT_BSE
+        incar_dict.pop('KPOINT_BSE', None)
+        #update T with True and F with False
+        incar_dict = {k: True if v == 'T' else False if v == 'F' else v for k, v in incar_dict.items()}
         return incar_dict
     
     def __getitem__(self, key: str) -> str:
@@ -50,8 +55,21 @@ class Incar:
             with open(path, 'w') as f:
                 f.write(self.__str__())
 
-    
 
+    def increment(self, increment_dict: dict[str, list], dry_run: bool = False):
+        '''Increments the INCAR file by a list of values for a given INCAR key'''
+
+        for key, values in increment_dict.items():
+            for value in values:
+                self.incar[key] = value
+                self.write_incar(f'INCAR_{key}_{value}', dry_run = dry_run)
+
+    
+    def to_yaml(self, path: Path):
+        '''Writes an identical INCAR file for the current calculation'''
+
+        with open(path, 'w') as f:
+            yaml.dump(self.incar, f)
